Helpers moved to shared module.
