+++ src/app/characters/pages/characters/characters.component.ts
@@ -1,7 +1,7 @@
 import {Component, OnDestroy, OnInit} from '@angular/core';
 import {CharactersService} from "../../characters.service";
 import {ActivatedRoute, Router} from "@angular/router";
-import {mapRouteParamsToCharacterQuery} from "../../helpers/charactersMapper";
+import {mapRouteParamsToCharacterQuery} from "../../../shared/helpers/charactersMapper";
 import {map, Subscription, tap} from "rxjs";
 
 @Component({
